Extracted Device config into its own component
To keep continuity across pages, a new component for the device config
was added. Same functionality, but no longer dependent on Manual
